% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross2rds.R
\name{cross2rds}
\alias{cross2rds}
\title{Read cross and save as rds}
\usage{
cross2rds(input_file, output_file, compress = FALSE)
}
\arguments{
\item{input_file}{Character string with path to the
\href{http://www.yaml.org}{YAML} or \href{http://www.json.org/}{JSON} file containing all of the control
information. This could instead be a zip file containing all of the
data files, in which case the contents are unzipped to a temporary
directory and then read.}

\item{output_file}{Character string with path to RDS file for output}

\item{compress}{If TRUE, save a compressed RDS file (smaller but slower).}
}
\description{
Read cross files and save as rds
}
\examples{
input_file <- paste0("https://github.com/rqtl/qtl2data/",
                     "blob/master/B6BTBR/b6btbr.zip")
\dontrun{cross2rds(input_file, "b6btbr.rds")}
}
