% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl2cl-package.R
\docType{package}
\name{qtl2cl-package}
\alias{qtl2cl}
\alias{qtl2cl-package}
\title{qtl2cl: Command-Line Interface for R/qtl2}
\description{
Support for a command-line interface for R/qtl2.
}
\section{Vignettes}{

\itemize{
\item \href{http://kbroman.org/qtl2/assets/vignettes/user_guide.html}{user guide}
\item \href{http://kbroman.org/qtl2/assets/vignettes/input_files.html}{input file formats}
\item \href{http://kbroman.org/qtl2/pages/prep_do_data.html}{preparing DO mouse data for R/qtl2}
\item \href{http://kbroman.org/qtl2/assets/vignettes/rqtl_diff.html}{differences between R/qtl and R/qtl2}
\item \href{http://kbroman.org/qtl2/assets/vignettes/developer_guide.html}{developer guide}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rqtl/qtl2cl}
}

}
\author{
\strong{Maintainer}: Karl W Broman \email{broman@wisc.edu} (0000-0002-4914-6671)

}
\keyword{internal}
