% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file.R
\name{read_file}
\alias{read_file}
\title{Read a file}
\usage{
read_file(file, ...)
}
\arguments{
\item{file}{Character string of input file path}

\item{...}{Passed to whatever function it calls}
}
\description{
Read a file that is either RDS, CSV, or JSON, with the method determined by the file extension.
}
\examples{
\dontrun{x <- read_file("myfile.rds")}
\dontrun{y <- read_file("myfile.csv")}
\dontrun{z <- read_file("myfile.json")}
}

