% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_calcgenoprob.R
\name{run_calcgenoprob}
\alias{run_calcgenoprob}
\title{Run calc_genoprob and save result to rds}
\usage{
run_calcgenoprob(cross_file, output_file, map_file = NULL, step = 0,
  off_end = 0, stepwidth = c("fixed", "max"), error_prob = 0.0001,
  map_function = c("haldane", "kosambi", "c-f", "morgan"), cores = 1,
  compress = FALSE)
}
\arguments{
\item{cross_file}{Character string with path to RDS file containing cross}

\item{output_file}{Character string with path to RDS file for output}

\item{map_file}{Character string with path to RDS file for writing genetic map
(with inserted pseudomarkers)}

\item{step}{Distance between pseudomarkers and markers; if
\code{step=0} no pseudomarkers are inserted.}

\item{off_end}{Distance beyond terminal markers in which to insert
pseudomarkers.}

\item{stepwidth}{Indicates whether to use a fixed grid
(\code{stepwidth="fixed"}) or to use the maximal distance between
pseudomarkers to ensure that no two adjacent markers/pseudomarkers
are more than \code{step} apart.}

\item{error_prob}{Assumed genotyping error probability}

\item{map_function}{Character string indicating the map function to
use to convert genetic distances to recombination fractions.}

\item{cores}{Number of CPU cores to use, for parallel calculations.
(If \code{0}, use \code{\link[parallel]{detectCores}}.)
Alternatively, this can be links to a set of cluster sockets, as
produced by \code{\link[parallel]{makeCluster}}.}

\item{compress}{If TRUE, save compressed RDS files (smaller but slower).}
}
\description{
Run calc_genoprob and save result to rds
}
\examples{
input_file <- paste0("https://github.com/rqtl/qtl2data/",
                     "blob/master/B6BTBR/b6btbr.zip")
\dontrun{cross2rds(input_file, "b6btbr.rds")}
\dontrun{run_calcgenoprob("b6btbr.rds", "b6btbr_probs.rds", "b6btbr_gmap.rds")}
}
