% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_calckinship.R
\name{run_calckinship}
\alias{run_calckinship}
\title{Calculate kinship matrix}
\usage{
run_calckinship(input_file, output_file, ...)
}
\arguments{
\item{input_file}{Input RDS file containing genotype or allele probabilities}

\item{output_file}{Output RDS file for calculated kinship matrix}

\item{...}{Passed to \code{\link[qtl2geno]{calc_kinship}}.}
}
\description{
Calculate genetic similarity among individuals (kinship matrix)
from conditional genotype probabilities.
}
\examples{
input_file <- paste0("https://github.com/rqtl/qtl2data/",
                     "blob/master/B6BTBR/b6btbr.zip")
\dontrun{cross2rds(input_file, "b6btbr.rds")}
\dontrun{run_calcgenoprob("b6btbr.rds", "b6btbr_probs.rds")}
\dontrun{run_calckinship("b6btbr_probs.rds", "b6btbr_kinship.rds")}
}

