% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_getXcovar.R
\name{run_getXcovar}
\alias{run_getXcovar}
\title{Get X covariates and write to file}
\usage{
run_getXcovar(input_file, output_file)
}
\arguments{
\item{input_file}{Input RDS file for cross}

\item{output_file}{Output RDS file for X chromosome covariates}
}
\description{
Read cross from RDS file, determine X chromosome covariates, and write to another RDS file.
}
\examples{
input_file <- paste0("https://github.com/rqtl/qtl2data/",
                     "blob/master/B6BTBR/b6btbr.zip")
\dontrun{cross2rds(input_file, "b6btbr.rds")}
\dontrun{run_getXcovar("b6btbr.rds", "b6btbr_xcovar.rds")}
}
