% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gp2ap.R
\name{run_gp2ap}
\alias{run_gp2ap}
\title{Run genoprob_to_alleleprob}
\usage{
run_gp2ap(input_file, output_file, cores = 1)
}
\arguments{
\item{input_file}{Name of input file (should be RDS)}

\item{output_file}{Name of output file (will be RDS)}

\item{cores}{Number of CPU cores to use, for parallel calculations.
(If \code{0}, use \code{\link[parallel]{detectCores}}.)
Alternatively, this can be links to a set of cluster sockets, as
produced by \code{\link[parallel]{makeCluster}}.}
}
\description{
Read in genotype probabilities, convert them to allele
probabilities, and write them back out.
}
\examples{
input_file <- paste0("https://github.com/rqtl/qtl2data/",
                     "blob/master/B6BTBR/b6btbr.zip")
\dontrun{cross2rds(input_file, "b6btbr.rds")}
\dontrun{run_calcgenoprob("b6btbr.rds", "b6btbr_probs.rds")}
\dontrun{run_gp2ap("b6btbr_probs.rds", "b6btbr_aprobs.rds")}
}
