% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scan1.R
\name{run_scan1}
\alias{run_scan1}
\title{Run scan1}
\usage{
run_scan1(genoprobs_file, pheno_file, output_file = NULL, map_file = NULL,
  kinship_file = NULL, addcovar_file = NULL, Xcovar_file = NULL,
  intcovar_file = NULL, weights_file = NULL, reml = TRUE, cores = 1,
  compress = FALSE)
}
\arguments{
\item{genoprobs_file}{Name of file with genotype probabilities}

\item{pheno_file}{Name of file with phenotypes}

\item{output_file}{Optional output RDS file. If NULL, print output as a table.}

\item{map_file}{Optional (RDS) file containing map. Needed if \code{output_file} is NULL.}

\item{kinship_file}{Optional file containing kinship matrix}

\item{addcovar_file}{Optional file containing additive covariates}

\item{Xcovar_file}{Optional file containing X chromosome covariates}

\item{intcovar_file}{Optional file containing interactive covariates}

\item{weights_file}{Optional file containing covariates}

\item{reml}{If TRUE, use REML; otherwise, use maximum likelihood}

\item{cores}{Number of CPU cores to use}

\item{compress}{If TRUE, save a compressed RDS file (smaller but slower).}
}
\description{
Read in a bunch of data and then run \code{\link[qtl2:scan1]{qtl2::scan1()}}.
}
