% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file.R
\name{read_file}
\alias{read_file}
\title{Read a file}
\usage{
read_file(file, ...)
}
\arguments{
\item{file}{Character string of input file; must have file
extension \code{.rds}, \code{.csv}, \code{.json}, or \code{.yaml}.}

\item{...}{Passed to whatever function it calls}
}
\value{
Could be most anything; whatever gets read in by the
corresponding R function for the file type. See Details.
}
\description{
Read a file that is either RDS, CSV, JSON, or YAML, with the method
determined by the file extension.
}
\details{
For CSV files, we use \code{\link[qtl2:read_csv]{qtl2::read_csv()}},
which calls \code{\link[data.table:fread]{data.table::fread()}} with a particular set
of options; note that columns are forced to be numeric in this
case.
For RDS files, we use \code{\link[base:readRDS]{base::readRDS()}}. For JSON files,
we use \code{\link[base:readLines]{base::readLines()}} and
\code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}}. For YAML files, we use
\code{\link[yaml:yaml.load_file]{yaml::yaml.load_file()}}.
}
\examples{
\dontrun{w <- read_file("myfile.rds")
x <- read_file("myfile.csv")
y <- read_file("myfile.json")
z <- read_file("myfile.yaml")}
}
