% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_calcgenoprob.R
\name{run_calcgenoprob}
\alias{run_calcgenoprob}
\title{Run calc_genoprob and save result to rds}
\usage{
run_calcgenoprob(cross_file, output_file, ...)
}
\arguments{
\item{cross_file}{Character string with path to RDS file containing cross}

\item{output_file}{Character string with path to RDS file for output}

\item{...}{Passed to \code{\link[qtl2geno]{calc_genoprob}}}
}
\value{
Name of output file
}
\description{
Run calc_genoprob and save result to rds
}
\examples{
input_file <- paste0("https://github.com/rqtl/qtl2data/",
                     "blob/master/B6BTBR/b6btbr.zip")
\dontrun{cross2rds(input_file, "b6btbr.rds")}
\dontrun{run_calcgenoprob("b6btbr.rds", "b6btbr_probs.rds")}
}

