% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_calckinship.R
\name{run_calckinship}
\alias{run_calckinship}
\title{Calculate kinship matrix}
\usage{
run_calckinship(input_file, output_file, type = c("overall", "loco", "chr"),
  omit_x = FALSE, use_allele_probs = TRUE, normalize = FALSE, cores = 1)
}
\arguments{
\item{input_file}{Input RDS file containing genotype or allele probabilities}

\item{output_file}{Output RDS file for calculated kinship matrix}

\item{type}{Indicates whether to calculate the overall kinship
(\code{"overall"}, using all chromosomes), the kinship matrix
leaving out one chromosome at a time (\code{"loco"}), or the
kinship matrix for each chromosome (\code{"chr"}).}

\item{omit_x}{If \code{TRUE}, only use the autosomes; ignored when
\code{type="chr"}.}

\item{use_allele_probs}{If \code{TRUE}, assess similarity with
allele probabilities (that is, first run
\code{\link{genoprob_to_alleleprob}}); otherwise use the genotype
probabilities.}

\item{normalize}{If \code{TRUE}, divide the kinship matrix by a
normalizing constant (see Details).}

\item{cores}{Number of CPU cores to use, for parallel calculations.
(If \code{0}, use \code{\link[parallel]{detectCores}}.)
Alternatively, this can be links to a set of cluster sockets, as
produced by \code{\link[parallel]{makeCluster}}.}
}
\description{
Calculate genetic similarity among individuals (kinship matrix)
from conditional genotype probabilities.
}
\examples{
input_file <- paste0("https://github.com/rqtl/qtl2data/",
                     "blob/master/B6BTBR/b6btbr.zip")
\dontrun{cross2rds(input_file, "b6btbr.rds")}
\dontrun{run_calcgenoprob("b6btbr.rds", "b6btbr_probs.rds")}
\dontrun{run_calckinship("b6btbr_probs.rds", "b6btbr_kinship.rds")}
}
